* 新增 vs 2015 的项目文件; * 修正大部分 x64 模式下的已知错误; * 新增 asmlib 的 64 位版本的库文件; * 还是基本上只能在 MSVC 里编译, 有1/3以上需要重构;

+++ src/jimi/lang/CharTraits.h
@@ -753,7 +753,7 @@ int char_traits<char_type>::utoa_radix10(char_type *buf, unsigned int val)
         *cur++ = static_cast<char_type>(digit_val + '0');
     } while (val != 0);
 
-    num_digits = cur - digits;
+    num_digits = (unsigned int)(cur - digits);
 
 #if 0
     do {
@@ -849,7 +849,7 @@ int char_traits<char_type>::ultoa_radix10(char_type *buf, unsigned long val)
         *cur++ = static_cast<char_type>(digit_val + '0');
     } while (val != 0);
 
-    num_digits = cur - digits;
+    num_digits = (unsigned long)(cur - digits);
 
 #if 0
     do {
